+++ packages/vite-plugin-vui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fastkit/vite-plugin-vui
 
+## 1.1.49
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fastkit/vui@0.16.46
+
 ## 1.1.48
 
 ### Patch Changes